# app/static/swagger.yaml
openapi: "3.0.3"
info:
  title: Velostadt API
  version: "1.0"
  description: REST API for the Velostadt e-scooter/e-bike rental platform.

servers:
  - url: /api

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: secret
        role:
          type: string
          enum: [benutzer, anbieter]
          default: benutzer

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
        user_id:
          type: integer

    Fahrzeug:
      type: object
      properties:
        fahrzeug_id:
          type: integer
        typ:
          type: string
        status:
          type: string
          enum: [verfuegbar, verliehen, in_wartung]
        akku_prozent:
          type: integer
        gps_lat:
          type: number
          nullable: true
        gps_long:
          type: number
          nullable: true
        qr_code:
          type: string
        basispreis:
          type: number
        preis_pro_min:
          type: number

    Fahrt:
      type: object
      properties:
        fahrt_id:
          type: integer
        fahrzeug_id:
          type: integer
        startzeit:
          type: string
          format: date-time
        endzeit:
          type: string
          format: date-time
          nullable: true
        kilometer:
          type: number
        preis:
          type: number

    Error:
      type: object
      properties:
        error:
          type: string

paths:
  /auth/login:
    post:
      summary: Erhalte ein JWT Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login erfolgreich
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Missing JSON body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /fahrzeuge:
    get:
      summary: List all vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [verfuegbar, verliehen, in_wartung]
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fahrzeug"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /fahrzeuge/{id}:
    get:
      summary: Get a single vehicle
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fahrzeug"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found

  /fahrten:
    get:
      summary: List rides
      description: >
        Benutzer see their own rides. Anbieter see rides on their vehicles.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of rides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fahrt"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"